requirements.txt
python-gitlab==3.3.0
prometheus-client==0.13.1
schedule==1.1.0
